requirements.txt
asn1crypto==0.22.0
bleach==2.0.0
cffi==1.10.0
cryptography==1.9
Django==1.11
django-ckeditor==5.2.2
django-colorful==1.2
django-debug-toolbar==1.8
django-jet==1.0.6
django-mptt==0.8.7
django-multiupload==0.5.2
django-recaptcha==1.3.0
django-sitetree==1.8.0
django-solo==1.1.2
easy-thumbnails==2.4.1
Fabric==1.13.2
html5lib==0.999999999
idna==2.5
olefile==0.44
paramiko==2.1.2
Pillow==4.1.1
psycopg2==2.7.1
pyasn1==0.2.3
pycparser==2.17
PyMySQL==0.7.11
pytz==2017.2
six==1.10.0
sqlparse==0.2.3
webencodings==0.5.1
